First steps executions
I had to solve a couple of problems regarding how my machien was (not) recognizign some commands.
One of them because is a Windows machine, meanign the environment cannot be activated via code from a Jupyter cell, specially not using the comand source. So, I had e to do this manually.
Another was that !pip was not working, so I changed it for %pip.
Changes in the code are commented.